Spinner (#213)
* Adds a spinner

* the specs
