+++ client/js/_author/components/assessments/question_types/multiple_choice/correct_selector.spec.jsx
@@ -1,0 +1,23 @@
+import React            from 'react';
+import { shallow }      from 'enzyme';
+import CorrectSelector  from './correct_selector';
+
+describe('Correct answer selector', () => {
+  let result;
+  let props;
+
+  beforeEach(() => {
+    props = {
+      itemType: '',
+      id: 'TacoId',
+      isCorrect: true,
+      updateChoice: () => {},
+    };
+    result = shallow(<CorrectSelector {...props} />);
+  });
+
+  it('checks if input.checked is true', () => {
+    expect(result.find('optionRadio').props().isCorrect).toBeTruthy();
+  });
+
+});
