+++ client/js/_author/components/assessments/question_types/multiple_choice/multiple_choice.spec.jsx
