+++ client/js/_author/components/assessments/question_types/multiple_choice/option_feedback.jsx
@@ -1,0 +1,24 @@
+import React    from 'react';
+import Editor   from '../../../common/oea_editor';
+
+export default function optionFeedback(props) {
+
+  return (
+    <div className="author--c-input author--c-input-label--left author--c-feedback">
+      <label htmlFor="feedback1">Feedback</label>
+      <Editor
+        text={props.feedback}
+        bankId={props.bankId}
+        itemId={props.itemId}
+        onBlur={(text, fileIds) => props.updateChoice({ feedback: text }, fileIds)}
+      />
+    </div>
+  );
+}
+
+optionFeedback.propTypes = {
+  feedback: React.PropTypes.string,
+  bankId: React.PropTypes.string.isRequired,
+  itemId: React.PropTypes.string.isRequired,
+  updateChoice: React.PropTypes.func.isRequired,
+};
